Upgrading to Spree 2.2
Former-commit-id: b3144e8d7c63d29ae4541cd4b459b86c1ca3fcba [formerly dc2ea49415a71d893ec281d349e60699cd35c713 [formerly 4a17a454c149b9f2f41b90e26f2a27b96d21ed23]]
Former-commit-id: d788c8f7df7a57428092321b87354430db091d27
Former-commit-id: c01d5581ca7ad82724d0965a4b08185398c2c80e [formerly b03c698a639defeb221c0dc4167d6de529eee012]
Former-commit-id: 6ec7909c616b750bae97ae1f7f2f9b38e01ebdf0

+++ core/app/models/spree/order.rb
@@ -5,14 +5,15 @@ require  File.join(Rails.root,'vendor/spree/core/app/jobs/spree/issue_gift_card_
 module Spree
   class Order < ActiveRecord::Base
     include Checkout
+    include CurrencyUpdater
     
     checkout_flow do
       go_to_state :address
       go_to_state :delivery
-      go_to_state :payment, if: ->(order) {
-        order.update_totals
+      go_to_state :payment, if: ->(order) do
+        order.set_shipments_cost if order.shipments.any?
         order.payment_required?
-      }
+      end
       go_to_state :confirm, if: ->(order) { order.confirmation_required? }
       go_to_state :complete
       remove_transition from: :delivery, to: :confirm
@@ -25,9 +26,11 @@ module Spree
     if Spree.user_class
       belongs_to :user, class_name: Spree.user_class.to_s
       belongs_to :created_by, class_name: Spree.user_class.to_s
+      belongs_to :approver, class_name: Spree.user_class.to_s
     else
       belongs_to :user, class_name: "Spree::User"
       belongs_to :created_by, class_name: "Spree::User"
+      belongs_to :approver
     end
 
     belongs_to :bill_address, foreign_key: :bill_address_id, class_name: 'Spree::Address'
@@ -36,16 +39,20 @@ module Spree
     belongs_to :ship_address, foreign_key: :ship_address_id, class_name: 'Spree::Address'
     alias_attribute :shipping_address, :ship_address
 
+    alias_attribute :ship_total, :shipment_total
+
     has_many :state_changes, as: :stateful
-    has_many :line_items, -> { order('created_at ASC') }, dependent: :destroy
+    has_many :line_items, -> { order('created_at ASC') }, dependent: :destroy, inverse_of: :order
     has_many :payments, dependent: :destroy
     has_many :return_authorizations, dependent: :destroy
     has_many :adjustments, -> { order("#{Adjustment.table_name}.created_at ASC") }, as: :adjustable, dependent: :destroy
     has_many :line_item_adjustments, through: :line_items, source: :adjustments
-    has_many :all_adjustments, class_name: 'Spree::Adjustment'
-    has_many :inventory_units
+    has_many :shipment_adjustments, through: :shipments, source: :adjustments
+    has_many :inventory_units, inverse_of: :order
 
-    has_many :shipments, dependent: :destroy, :class_name => "Shipment" do
+    has_and_belongs_to_many :promotions, join_table: 'spree_orders_promotions'
+
+    has_many :shipments, dependent: :destroy, inverse_of: :order do
       def states
         pluck(:state).uniq
       end
@@ -69,14 +76,15 @@ module Spree
     attr_accessor :use_billing
 
     before_create :link_by_email
-    after_create :create_tax_charge!
+    before_update :homogenize_line_item_currencies, if: :currency_changed?
 
     validates :email, presence: true, if: :require_email
     validates :email, email: true, if: :require_email, allow_blank: true
     validate :has_available_shipment
-    validate :has_available_payment
 
     make_permalink field: :number
+
+    delegate :update_totals, :persist_totals, :to => :updater
 
     class_attribute :update_hooks
     self.update_hooks = Set.new
@@ -142,10 +150,6 @@ module Spree
       parcels_grouped_by_box.map(&:longest_edge).sort{ |a,b| b <=> a }.first
     end
 
-    def self.complete
-      where.not(completed_at: nil)
-    end
-
 
     def metapack_booking_code
       li_by_product_type = line_items.map {|li| li.variant.product.product_type == 'pattern'}
@@ -154,37 +158,6 @@ module Spree
       ((has_only_pattern && less_than_ten) ? 'PATTERN' : shipping_zone_name)
     end
 
-    def shipping_zone_name
-      country_id = (self.ship_address.try(:country).nil? ? nil : self.ship_address.try(:country).id)
-      if country_id
-        zone = Spree::Zone.
-          includes(:zone_members).
-          where("spree_zone_members.zoneable_id = ? AND spree_zone_members.zoneable_type = ?", country_id, "Spree::Country").
-          reorder('zone_members_count', 'spree_zones.created_at').
-          references(:zone_members).
-          first
-        zone.name.upcase if zone
-      end
-    end
-
-    def parcels_grouped_by_box
-      parcels.inject([]) do |list, parcel|
-        current_parcel = list.detect {|e| e.box_id == parcel.box_id}
-        if current_parcel
-          current_parcel.quantity += 1
-        else
-          current_parcel = parcel
-          current_parcel.quantity = 1
-          list << current_parcel
-        end
-
-        list
-      end
-    end
-
-    def self.incomplete
-      where(completed_at: nil)
-    end
 
     def item_normal_total
       if @item_normal_total.blank?
@@ -194,10 +167,8 @@ module Spree
       end
     end
 
-    # Use this method in other gems that wish to register their own custom logic
-    # that should be called after Order#update
-    def self.register_update_hook(hook)
-      self.update_hooks.add(hook)
+    def all_adjustments
+      Adjustment.where("order_id = :order_id OR adjustable_id = :order_id", :order_id => self.id)
     end
 
     # For compatiblity with Calculator::PriceSack
@@ -236,16 +207,29 @@ module Spree
       Spree::Money.new(adjustment_total, { currency: currency })
     end
 
-    def display_tax_total
-      Spree::Money.new(tax_total, { currency: currency })
+    def display_included_tax_total
+      Spree::Money.new(included_tax_total, { currency: currency })
     end
 
-    def display_ship_total
-      Spree::Money.new(ship_total, { currency: currency })
+    def display_additional_tax_total
+      Spree::Money.new(additional_tax_total, { currency: currency })
     end
+
+    def display_tax_total
+      Spree::Money.new(included_tax_total + additional_tax_total, { currency: currency })
+    end
+
+    def display_shipment_total
+      Spree::Money.new(shipment_total, { currency: currency })
+    end
+    alias :display_ship_total :display_shipment_total
 
     def display_total
       Spree::Money.new(total, { currency: currency })
+    end
+
+    def shipping_discount
+      shipment_adjustments.eligible.sum(:amount) * - 1
     end
 
     def to_param
@@ -286,11 +270,6 @@ module Spree
         state == 'confirm'
     end
 
-    # Indicates the number of items in the order
-    def item_count
-      line_items.inject(0) { |sum, li| sum + li.quantity }
-    end
-
     def backordered?
       shipments.any?(&:backordered?)
     end
@@ -314,26 +293,12 @@ module Spree
       Spree::Config[:tax_using_ship_address] ? ship_address : bill_address
     end
 
-    # Array of totals grouped by Adjustment#label. Useful for displaying line item
-    # adjustments on an invoice. For example, you can display tax breakout for
-    # cases where tax is included in price.
-    def line_item_adjustment_totals
-      Hash[self.line_item_adjustments.eligible.group_by(&:label).map do |label, adjustments|
-        total = adjustments.sum(&:amount)
-        [label, Spree::Money.new(total, { currency: currency })]
-      end]
-    end
-
     def updater
       @updater ||= OrderUpdater.new(self)
     end
 
     def update!
       updater.update
-    end
-
-    def update_totals
-      updater.update_totals
     end
 
     def clone_billing_address
@@ -398,14 +363,11 @@ module Spree
       line_items.detect { |line_item| line_item.variant_id == variant.id }
     end
 
-    def ship_total
-      adjustments.shipping.sum(:amount)
-    end
-
     # Creates new tax charges if there are any applicable rates. If prices already
     # include taxes then price adjustments are created instead.
     def create_tax_charge!
-      Spree::TaxRate.adjust(self)
+      Spree::TaxRate.adjust(self, line_items)
+      Spree::TaxRate.adjust(self, shipments) if shipments.any?
     end
 
     def outstanding_balance
@@ -413,7 +375,7 @@ module Spree
     end
 
     def outstanding_balance?
-      self.outstanding_balance != 0
+     self.outstanding_balance != 0
     end
 
     def name
@@ -434,10 +396,8 @@ module Spree
     # Finalizes an in progress order after checkout is complete.
     # Called after transition to complete state when payments will have been processed
     def finalize!
-      touch :completed_at
-
       # lock all adjustments (coupon promotions, etc.)
-      adjustments.update_all state: 'closed'
+      all_adjustments.each{|a| a.close}
 
       # update payment and shipment(s) states, and save
       updater.update_payment_state
@@ -450,11 +410,15 @@ module Spree
       save
       updater.run_hooks
 
-      deliver_order_confirmation_email
-      deliver_gift_card_emails
-    end
+      touch :completed_at
+      
+	  deliver_gift_card_emails    
+      deliver_order_confirmation_email unless confirmation_delivered?
 
-    def deliver_gift_card_emails
+      consider_risk
+	end
+    
+	def deliver_gift_card_emails
       self.gift_card_line_items.each do |item|
         item.quantity.times {|position| 
           job = Spree::IssueGiftCardJob.new(self, item, position)
@@ -486,6 +450,7 @@ module Spree
     def deliver_order_confirmation_email
       begin
         OrderMailer.confirm_email(self.id).deliver
+		update_column(:confirmation_delivered, true)
       rescue Exception => e
         logger.error("#{e.class.name}: #{e.message}")
         logger.error(e.backtrace * "\n")
@@ -502,7 +467,7 @@ module Spree
     end
 
     def pending_payments
-      payments.select(&:checkout?)
+      payments.select { |payment| payment.checkout? || payment.pending? }
     end
 
     # processes any pending payments and must return a boolean as it's
@@ -584,13 +549,12 @@ module Spree
     end
 
     def empty!
-      adjustments.destroy_all
       line_items.destroy_all
-    end
+      updater.update_item_count
 
-    def clear_adjustments!
-      self.adjustments.destroy_all
-      self.line_item_adjustments.destroy_all
+      adjustments.destroy_all
+      update_totals
+      persist_totals
     end
 
     def has_step?(step)
@@ -617,27 +581,10 @@ module Spree
       @coupon_code = code.strip.downcase rescue nil
     end
 
-    # Tells us if there if the specified promotion is already associated with the order
-    # regardless of whether or not its currently eligible. Useful because generally
-    # you would only want a promotion action to apply to order no more than once.
-    #
-    # Receives an adjustment +originator+ (here a PromotionAction object) and tells
-    # if the order has adjustments from that already
-    def promotion_credit_exists?(originator)
-      !! adjustments.includes(:originator).promotion.reload.detect { |credit| credit.originator.id == originator.id }
+    def can_add_coupon?
+      Spree::Promotion.order_activatable?(self)
     end
 
-    def promo_total
-      adjustments.eligible.promotion.sum(:amount)
-    end
-
-    def manual_adjustment_total
-      adjustments.eligible.manual.sum(:amount)
-    end
-
-    def discount_total
-      promo_total + manual_adjustment_total
-    end
 
     def shipped?
       %w(partial shipped).include?(shipment_state)
@@ -655,42 +602,102 @@ module Spree
       shipments
     end
 
+    def apply_free_shipping_promotions
+      Spree::PromotionHandler::FreeShipping.new(self).activate
+      shipments.each { |shipment| ItemAdjustments.new(shipment).update }
+      updater.update_shipment_total
+      persist_totals
+    end
+
     # Clean shipments and make order back to address state
     #
     # At some point the might need to force the order to transition from address
     # to delivery again so that proper updated shipments are created.
     # e.g. customer goes back from payment step and changes order items
     def ensure_updated_shipments
-      if shipments.any?
+      if shipments.any? && !self.completed?
         self.shipments.destroy_all
+        self.update_column(:shipment_total, 0)
         restart_checkout_flow
       end
     end
 
     def restart_checkout_flow
-      self.update_column(:state, checkout_steps.first)
+      self.update_columns(
+        state: checkout_steps.first,
+        updated_at: Time.now,
+      )
     end
 
     def refresh_shipment_rates
-      shipments.map(&:refresh_rates)
+      shipments.map &:refresh_rates
     end
 
     def shipping_eq_billing_address?
       (bill_address.empty? && ship_address.empty?) || bill_address.same_as?(ship_address)
     end
 
-    def can_attempt_payment?
-      payments.select(&:pending?).blank?
+    def set_shipments_cost
+      shipments.each(&:update_amounts)
+      updater.update_shipment_total
+      persist_totals
     end
 
     def is_risky?
       self.payments.where(%{
-        (avs_response IS NOT NULL and avs_response != 'D') or
+        (avs_response IS NOT NULL and avs_response != '' and avs_response != 'D' and avs_response != 'M') or
         (cvv_response_code IS NOT NULL and cvv_response_code != 'M') or
-        cvv_response_message IS NOT NULL or
+        cvv_response_message IS NOT NULL and cvv_response_message != '' or
         state = 'failed'
-        }.squish!).uniq.count > 0
+      }.squish!).uniq.count > 0
     end
+
+    def approved_by(user)
+      self.transaction do
+        approve!
+        self.update_columns(
+          approver_id: user.id,
+          approved_at: Time.now,
+          considered_risky: false,
+        )
+      end
+    end
+
+    def approved?
+      !!self.approved_at
+    end
+
+    def can_approve?
+      !approved?
+    end
+
+    def consider_risk
+      if is_risky? && !approved?
+        considered_risky!
+      end
+    end
+
+    def considered_risky!
+      update_column(:considered_risky, true)
+    end
+
+    def approve!
+      update_column(:considered_risky, false)
+    end
+
+    # moved from api order_decorator. This is a better place for it.
+    def update_line_items(line_item_params)
+      return if line_item_params.blank?
+      line_item_params.each_value do |attributes|
+        if attributes[:id].present?
+          self.line_items.find(attributes[:id]).update_attributes!(attributes)
+        else
+          self.line_items.create!(attributes)
+        end
+      end
+      self.ensure_updated_shipments
+    end
+
 
     def find_existing_line_item(variant, parts, personalisations, target_id)
       uuid = Spree::VariantUuid.fetch(variant, parts, personalisations).number
@@ -727,33 +734,30 @@ module Spree
       end
     end
 
-    def has_available_payment
-      return unless delivery?
-      # errors.add(:base, :no_payment_methods_available) if available_payment_methods.empty?
-    end
+      def after_cancel
+        shipments.each { |shipment| shipment.cancel! }
+        payments.completed.each { |payment| payment.credit! }
 
-    def after_cancel
-      shipments.each { |shipment| shipment.cancel! }
-      payments.completed.each { |payment| payment.credit! }
+        send_cancel_email
+        self.update_column(:payment_state, 'credit_owed') unless shipped?
+      end
 
-      send_cancel_email
-      self.update_column(:payment_state, 'credit_owed') unless shipped?
-    end
+      def send_cancel_email
+        OrderMailer.cancel_email(self.id).deliver
+      end
 
-    def send_cancel_email
-      OrderMailer.cancel_email(self.id).deliver
-    end
+      def after_resume
+        shipments.each { |shipment| shipment.resume! }
+        consider_risk
+      end
 
-    def after_resume
-      shipments.each { |shipment| shipment.resume! }
-    end
+      def use_billing?
+        @use_billing == true || @use_billing == 'true' || @use_billing == '1'
+      end
 
-    def use_billing?
-      @use_billing == true || @use_billing == 'true' || @use_billing == '1'
-    end
+      def set_currency
+        self.currency = Spree::Config[:currency] if self[:currency].nil?
+      end
 
-    def set_currency
-      self.currency = Spree::Config[:currency] if self[:currency].nil?
-    end
   end
 end
